refactor(herdr): agent.rs / state.rs の Value 手掘りを生成型へ移す (#535)
* refactor(herdr): agent.rs / state.rs の Value 手掘りを生成型へ移す

読む側 6 メソッド（pane.get / pane.list / workspace.list /
workspace.create / pane.read / agent.start）と、送る側 22 メソッドの
params を生成型に通す。transport に call_typed / to_params を足した。

実行時の振る舞いは変えない。型化が変えるのは「どこで気づくか」だけで、
「気づいた後どうするか」は呼び出しごとに既に設計されている。

型化で分かったこと 3 つ:

1. FakeHerdr が実 herdr と乖離していた（workspace.create の root_pane に
   agent_status / focused / revision / terminal_id が無い、WorkspaceInfo
   の 8 必須のうち 6 つが無い、AgentSessionInfo の 4 必須のうち 3 つが
   無い等）。実 herdr が送れない形が全テストを緑で通していた。
   fixture ビルダを生成型で検査するテストを足して再発を止める
2. workspace.create の root_pane を required のまま読むと、欠落時に
   NewWorkspace が出来る前に落ち、作った workspace を閉じられず漏らす。
   そこだけ Option に緩めた CreatedWorkspace を置く（後始末の順序の
   問題であって形の問題ではない）
3. events.subscribe の pane.exited は schema 上ユニットバリアントで
   pane_id を取らない。購読はグローバルで、送っている pane_id は herdr が
   無視するキー。それでも送るのは transport::subscribed_panes が totsuka
   自身のために読み返しているため

生成器も 2 点直した: `format: float` を f32 でなく f64 へ写す（f32 だと
設定の 0.65 が 0.6499999761581421 として送られる）／`skip_serializing_if`
を Option のときだけに絞る（空のコレクションは値である）。

config/validate は ping に加えて workspace.list を 1 回型でパースする。
「この herdr の応答を読めない」をタスク取り込み前に doctor が言える。

テストは 120 → 124（削除ゼロ、追加 4: 送れない形の拒否 / fixture の
schema 適合 / config validate の 2 経路）。

refs #519, #517


* fix(herdr): 型化で失いかけた degrade を戻し、主張を実装に合わせる

/code-review の指摘 7 件への対応。上位 2 件は「振る舞いは変えていない」と
いう主張そのものに当たっていた。

1. CreatedWorkspace の緩和が漏れの一部しか塞いでいなかった。`root_pane`
   が「在るのに壊れている」場合と、`workspace` が workspace_id 以外の
   必須 7 つのどれかを欠く場合は、id が応答に在るのに掃除できずに落ちる。
   使うフィールドだけを読む形（CreatedWorkspaceRef / CreatedPaneRef）へ
   狭めた。これは設計の穴ではなく「実行時は寛容・CI は厳格」の実践で、
   ここで厳格にしても早く気づけず workspace が漏れるだけ
2. session/list が #416 の degrade（読めない応答は「何も所有していない」）
   を失っていた。doctor の孤児 pane 検出が乗る best-effort な診断面なので
   明示的に復活させた（unreadable_is_empty）。呼び出し自体の失敗は従来
   どおり伝播する。session/attach のほうは意図的にエラーへ変えた側なので、
   「一切変えていない」という書き方をやめて両方を名指しした
3. 拒否テストが名前の主張を固定していなかった（label 以外も欠けていたので
   Option<String> に戻しても緑だった）。label だけを動かす形に直した
4. 手前に挿入したせいで置き去りになった doc 2 箇所と、移設元に残っていた
   「32 slot」の説明を掃除
5. to_params の doc が実装より強かった（6 箇所は unwrap_or(Value::Null)）。
   2 種類の呼び出し側があることを doc に書いた
6. herdr-types.jq の関数 doc が「キーごと落とす」と無条件に述べたまま
   だった（この PR で Option 限定に狭めた）
7. component doc の server 行が config/validate の変更に触れていなかった

テストは 124 → 126（malformed root_pane の後始末 / session/list の degrade）。
どちらも無効化して落ちることを確認した。

refs #519, #517


* fix(herdr): workspace.create を生の Value 読みに戻し、agent.start を型で受ける

Copilot レビューの指摘 4 件への対応。

1. `Option<CreatedPaneRef>` は `root_pane` が「在るが型が違う」（`{"pane_id": 1}`）
   ときも全体のデシリアライズを落とす — **serde は Option の内側のエラーを
   伝播する（飲み込まない）**。緩め方を 2 回間違えていた。落ちない形は
   「serde にエラーで答えられる問いを投げない形」＝生の Value から 2 つの
   id を拾うことだけだった。生成型は herdr の約束の記録として残す
2. `agent.start` の成功応答が Value のままで、生成した AgentStartedEnvelope
   は production から一度も使われていなかった（PR 本文の「読む側 6 メソッド
   を型化」に反する）。成功だけ型で受け、リトライ判断に使うエラー経路は
   生のまま残す
3. config/validate の型パース失敗が「どの herdr が」を落としていた。
   ping で得た version を保持して添える
4. fixture の schema 適合検査が AgentInfo と PaneReadResult を見ていな
   かった（AgentInfo は PaneInfo と別の型、pane.read はビルダを通らない）

テストは `malformed_root_pane` を「型違いの pane_id」へ強め、workspace が
漏れないことを固定した。狭めた版に戻すと落ちることを確認済み。

refs #519, #517


* docs(herdr): apply_layout の env 説明を、実際に送っている形へ直す

「simply not passing `env` is what removes it」は #519 以降そうではない。
生成した `PaneSplitParams.env` は `Option` ではなく map なので、未設定の
env は `"env": {}` として wire に載る。適用される変数はどちらもゼロだが、
シェル pane にフックトークンが載らない理由は**キーの不在ではなく
`pane.split` が継承しないこと**なので、そう書き直す。

セキュリティ前提の説明が、それを支えている性質と別のことを言っている
状態だった（レビューで差分外として指摘）。

refs #519, #517

+++ plugins/agent-ide-herdr/src/wire.rs
@@ -43,7 +43,7 @@ pub mod request {
     pub struct AgentPromptWaitOptions {
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub timeout_ms: Option<u64>,
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
+        #[serde(default)]
         pub until: Vec<AgentStatus>,
     }
 
@@ -55,7 +55,7 @@ pub mod request {
 
     #[derive(Debug, Clone, Serialize)]
     pub struct AgentStartParams {
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
+        #[serde(default)]
         pub args: Vec<String>,
         pub kind: String,
         pub name: String,
@@ -83,7 +83,7 @@ pub mod request {
         pub target: String,
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub timeout_ms: Option<u64>,
-        #[serde(default, skip_serializing_if = "Vec::is_empty")]
+        #[serde(default)]
         pub until: Vec<AgentStatus>,
     }
 
@@ -147,11 +147,11 @@ pub mod request {
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub seq: Option<u64>,
         pub source: String,
-        #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
+        #[serde(default)]
         pub state_labels: BTreeMap<String, String>,
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub title: Option<String>,
-        #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
+        #[serde(default)]
         pub tokens: BTreeMap<String, Option<String>>,
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub ttl_ms: Option<u64>,
@@ -168,12 +168,12 @@ pub mod request {
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub cwd: Option<String>,
         pub direction: SplitDirection,
-        #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
+        #[serde(default)]
         pub env: BTreeMap<String, String>,
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub focus: Option<bool>,
         #[serde(default, skip_serializing_if = "Option::is_none")]
-        pub ratio: Option<f32>,
+        pub ratio: Option<f64>,
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub target_pane_id: Option<String>,
         #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -294,7 +294,7 @@ pub mod request {
     pub struct WorkspaceCreateParams {
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub cwd: Option<String>,
-        #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
+        #[serde(default)]
         pub env: BTreeMap<String, String>,
         #[serde(default, skip_serializing_if = "Option::is_none")]
         pub focus: Option<bool>,
